+++ examples/014-pc-uart-comm/f407zg/inc/config.h
@@ -30,6 +30,9 @@ extern "C" {
 #define _USE_USART1_PA9PA10
 #include "uart/uart.h"
 
+// ----------------------------------------------------------------------------
+#include "protocol.h"
+
 // ============================================================================
 #ifdef __cplusplus
 }
