+++ examples/014-pc-uart-comm/f767zi/src/main.c
@@ -1,6 +1,6 @@
 #include "config.h"
 
-#include "comm-protocol.h"
+#include "protocol.h"
 #include <math.h>
 
 uint8_t usart2_rx[100];
