+++ examples/014-pc-uart-comm/pc-program/comm/comm.cpp
@@ -1,5 +1,4 @@
-#include "serial-comm.h"
-#include "comm-protocol.h"
+#include "comm.h"
 #include <iostream>
 
 PcComm::PcComm(Serial* serial) {
